requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
asttokens==2.2.1
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
backcall==0.2.0
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
datetime==4.9
decorator==5.1.1 ; python_version >= '3.5'
executing==1.2.0
frozenlist==1.3.3 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
ipython==8.13.1 ; python_version >= '3.9'
jedi==0.18.2 ; python_version >= '3.6'
json5==0.9.11
matplotlib-inline==0.1.6 ; python_version >= '3.5'
multidict==6.0.4 ; python_version >= '3.7'
numpy==1.24.3 ; python_version >= '3.10'
packaging==23.1 ; python_version >= '3.7'
pandas==1.5.3 ; python_version >= '3.8'
parso==0.8.3 ; python_version >= '3.6'
pexpect==4.8.0 ; sys_platform != 'win32'
pickleshare==0.7.5
plotly==5.14.1
prompt-toolkit==3.0.38 ; python_full_version >= '3.7.0'
psutil==5.9.5 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.15.1 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
setuptools==67.7.2 ; python_version >= '3.7'
shimoku-api-python==0.19
shimoku-components-catalog==0.2.1
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
stack-data==0.6.2
tenacity==8.1.0 ; python_version >= '3.6'
tqdm==4.64.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
traitlets==5.9.0 ; python_version >= '3.7'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
wcwidth==0.2.6
yarl==1.9.2 ; python_version >= '3.7'
zope.interface==6.0 ; python_version >= '3.7'

pandas~=1.5.3